+++ src/SystemMonitor.cpp
@@ -1,12 +1,12 @@
 // Copyright (c) 2018-2026 Smart Swimming Pool, Stephan Strittmatter
 
-#include "SystemMonitor.hpp"
+#include "src/SystemMonitor.hpp"
 
 namespace PoolController {
 
 // Static member initialization
-unsigned long SystemMonitor::lastMemoryCheck  = 0;
-uint32_t      SystemMonitor::minFreeHeap      = 0;
-bool          SystemMonitor::lowMemoryWarning = false;
+uint32_t SystemMonitor::lastMemoryCheck = 0;
+uint32_t SystemMonitor::minFreeHeap = 0;
+bool SystemMonitor::lowMemoryWarning = false;
 
 }  // namespace PoolController
